5/12/2017 - Edited Menu Views

+++ fishingChampion/src/byui/cit260/fishingChampion/model/Map.java
@@ -13,45 +13,45 @@ import java.io.Serializable;
  */
 public class Map implements Serializable{
     //class instance variables.
-    private double rowCount;
-    private double columnCount;
-    private double currentRow;
-    private double currentColumn;
+    private int rowCount;
+    private int columnCount;
+    private int currentRow;
+    private int currentColumn;
     private double currentScene;
 
     public Map(int i, int i0) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
 
-    public double getRowCount() {
+    public int getRowCount() {
         return rowCount;
     }
 
-    public void setRowCount(double rowCount) {
+    public void setRowCount(int rowCount) {
         this.rowCount = rowCount;
     }
 
-    public double getColumnCount() {
+    public int getColumnCount() {
         return columnCount;
     }
 
-    public void setColumnCount(double columnCount) {
+    public void setColumnCount(int columnCount) {
         this.columnCount = columnCount;
     }
 
-    public double getCurrentRow() {
+    public int getCurrentRow() {
         return currentRow;
     }
 
-    public void setCurrentRow(double currentRow) {
+    public void setCurrentRow(int currentRow) {
         this.currentRow = currentRow;
     }
 
-    public double getCurrentColumn() {
+    public int getCurrentColumn() {
         return currentColumn;
     }
 
-    public void setCurrentColumn(double currentColumn) {
+    public void setCurrentColumn(int currentColumn) {
         this.currentColumn = currentColumn;
     }
